smtplib: Get struct sockaddr_un from caller
As discussed in micropython/micropython#1375, the Unix port's
usocket.connect method takes a bytes object that represents a
struct sockaddr. In the case of Unix domain sockets, it's a
struct sockaddr_un. There's no portable way to pack that struct (since
it can be defined differently on different platforms), so we'll let the
caller handle it. On many Linux platforms, this can be achieved with
something like ustruct.pack('H108s', usocket.AF_UNIX, '/path/to/sock').