smtplib: Prefer self.file over self.sock
The result of MicroPython's ussl.wrap_socket does not have the send or
makefile methods, unlike CPython. Because of the possibility that
starttls has been called (not to mention the possibility that SSL was
used from the outset, as in the SMTP_SSL class), it is not generally
known whether self.sock is an SSL-wrapped socket or not. As such, we
cannot, in general, call self.sock.makefile or self.sock.send.

Instead of potentially calling self.sock.makefile in getreply (as is
done upstream), I've opted to create self.file exactly when creating
self.sock (no more or less). This works because, when creating
self.sock, it is always known whether self.sock is SSL-wrapped or not,
and thus it is known how to get a self.file.

An alternative solution (for MicroPython) would be to never call
self.sock.makefile, and take advantage of the fact that MicroPython
sockets have read and write methods already (using write instead of
send). That seemed less desirable though, since it would potentially
make merging of upstream patches difficult, while providing no clear
benefit. (The two options seemed like a wash in terms of minimizing the
size of the diff from upstream.)